backend/requirements.txt
# web framework
fastapi
uvicorn
# database
tortoise-orm
# models
pydantic[dotenv]
# json accelerator (optional)
python-rapidjson
